[Plugin Hub] Refactor Single Page Generation
Pull the information from the following files:
 * _configuration.yml
 * _index.md
 * _changelog.md

and build the same page.

The most basic file structure for a plugin is:
app/
  _hub/
    kong-inc/
      /plugin
        _changelog.md
        _configuration.yml
        _index.md

The content and the configuration parameters table can be overriden for
specific versions by adding a folder with the corresponding version
number, i.e.

app/
  _hub/
    kong-inc/
      plugin/
        _changelog.md
        _configuration.yml
        _index.md
        _2.2.x/
          _configuration.yml
          _index.md

More granular changes can be made with the single-source conditional
rendering helpers.
Note: We use just one changelog file for all the versions.